analysis_nas {creditmodel}	R Documentation

Missing Analysis

Description

#' analysis_nas is for understanding the reason for missing data and understand distribution of missing data so we can categorise it as:

Missing completely at random(MCAR)

MMissing at random(MAR), or

Missing not at random, also known as IM.

Usage

analysis_nas(dat, class_var = FALSE, nas_rate = NULL, na_vars = NULL,
  mat_nas_shadow = NULL, dt_nas_random = NULL, ...)

Arguments

dat
A data.frame with independent variables and target variable.

class_var
Logical, nas analysis of the nominal variables. Default is TRUE.

nas_rate
A list contains nas rate of each variable.

na_vars
Names of variables which contain nas.

mat_nas_shadow
A shadow matrix of variables which contain nas.

dt_nas_random
A data.frame with random nas imputation.

...
Other parameters.

Value

A data.frame with outliers analysis for each variable.

[Package creditmodel version 1.1.2 Index]