auc_value {creditmodel}	R Documentation

auc_value auc_value is for get best lambda required in lasso_filter. This function required in lasso_filter

Description

auc_value auc_value is for get best lambda required in lasso_filter. This function required in lasso_filter

Usage

auc_value(target, prob)

Arguments

target
Vector of target.

prob
A list of redict probability or score.

Value

Lanmbda value

[Package creditmodel version 1.1.2 Index]