char_cor_vars {creditmodel}	R Documentation

Cramer's V matrix between categorical variables.

Description

char_cor_vars is function for calculating Cramer's V matrix between categorical variables. char_cor is function for calculating the correlation coefficient between variables by cremers 'V

Usage

char_cor_vars(dat, x)

char_cor(dat, x_list = NULL, ex_cols = "date$", parallel = FALSE,
  note = FALSE)

Arguments

dat
A data frame.

x
The name of variable to process.

x_list
Names of independent variables.

ex_cols
A list of excluded variables. Regular expressions can also be used to match variable names. Default is NULL.

parallel
Logical, parallel computing. Default is FALSE.

note
Logical. Outputs info. Default is TRUE.

Value

A list contains correlation index of x with other variables in dat.

Examples

## Not run: 
char_x_list = get_names(dat = UCICreditCard,
types = c('factor', 'character'),
ex_cols = "ID$|date$|default.payment.next.month$", get_ex = FALSE)
 char_cor(dat = UCICreditCard[char_x_list])

## End(Not run)

[Package creditmodel version 1.1.2 Index]