checking_data {creditmodel}	R Documentation

Checking Data

Description

checking_data cheking dat before processing.

Usage

checking_data(dat = NULL, target = NULL, occur_time = NULL,
  note = FALSE, pos_flag = NULL)

Arguments

dat
A data.frame with independent variables and target variable.

target
The name of target variable. Default is NULL.

occur_time
The name of the variable that represents the time at which each observation takes place.

note
Logical.Outputs info.Default is TRUE.

pos_flag
The value of positive class of target variable, default: "1".

Value

data.frame

Examples

dat = checking_data(dat = UCICreditCard, target = "default.payment.next.month")

[Package creditmodel version 1.1.2 Index]