data_exploration {creditmodel}	R Documentation

Data Exploration

Description

#'The data_exploration includes both univariate and bivariate analysis and ranges from univariate statistics and frequency distributions, to correlations, cross-tabulation and characteristic analysis.

Usage

data_exploration(dat, target = NULL, save_data = FALSE,
  file_name = NULL, dir_path = tempdir())

Arguments

dat
A data.frame with x and target.

target
The name of target variable.

save_data
Logical. If TRUE, save files to the specified folder at dir_path

file_name
The file name for periodically saved outliers analysis file. Default is NULL.

dir_path
The path for periodically saved outliers analysis file. Default is tempdir().

Value

A list contains both categrory and numeric variable analysis.

Examples

data_ex = data_exploration(dat = lendingclub[1:1000,],target = "loan_status")

[Package creditmodel version 1.1.2 Index]