derived_ts_vars {creditmodel}	R Documentation

Derivation of Behavioral Variables

Description

This function is used for derivating behavioral variables and is not intended to be used by end user.

Usage

derived_ts_vars(dat, grx, td = 12, der = c("cvs", "sums", "means",
  "maxs", "max_mins", "time_intervals", "cnt_intervals", "total_pcts",
  "cum_pcts", "partial_acfs"), parallel = TRUE)

derived_ts(dat = dat, grx_x = NULL, td = 12, der = c("cvs", "sums",
  "means", "maxs", "max_mins", "time_intervals", "cnt_intervals",
  "total_pcts", "cum_pcts", "partial_acfs"))

Arguments

dat
A data.frame contained only predict variables.

grx
Regular expressions used to match variable names.

td
Number of variables to derivate.

der
Variables to derivate

parallel
Logical, parallel computing. Default is FALSE.

grx_x
Regular expression used to match a group of variable names.

Details

The key to creating a good model is not the power of a specific modelling technique, but the breadth and depth of derived variables that represent a higher level of knowledge about the phenomena under examination.

[Package creditmodel version 1.1.2 Index]