entry_rate_na {creditmodel}	R Documentation

Max Percent of Missing Value

Description

entry_rate_na is the function to recode variables which NAs or miss value rate is more than 95

Usage

entry_rate_na(dat, nr = 0.95, note = FALSE)

Arguments

dat
A data frame with x and target.

nr
The maximum percent of NAs.

note
Logical.Outputs info.Default is TRUE.

Value

A data.frame

Examples

datss = entry_rate_na(dat = lendingclub[1:1000, ], nr = 0.98)

[Package creditmodel version 1.1.2 Index]