get_auc_ks_lambda {creditmodel}	R Documentation

get_auc_ks_lambda get_auc_ks_lambda is for get best lambda required in lasso_filter. This function required in lasso_filter

Description

get_auc_ks_lambda get_auc_ks_lambda is for get best lambda required in lasso_filter. This function required in lasso_filter

Usage

get_auc_ks_lambda(lasso_model, x_test, y_test, save_data = FALSE,
  plot_show = TRUE, file_name = NULL, dir_path = tempdir())

Arguments

lasso_model
A lasso model genereted by glmnet.

x_test
A matrix of test dataset with x.

y_test
A matrix of y test dataset with y.

save_data
Logical, save results in locally specified folder. Default is FALSE

plot_show
Logical, if TRUE plot the results. Default is TRUE.

file_name
The name for periodically saved results files. Default is NULL.

dir_path
The path for periodically saved results files.

Value

Lanmbda values with max K-S and AUC.

See Also

lasso_filter, get_sim_sign_lambda

[Package creditmodel version 1.1.2 Index]