get_correlation_group {creditmodel}	R Documentation

get_correlation_group

Description

get_correlation_group is funtion for obtaining highly correlated variable groups. select_cor_group is funtion for selecting highly correlated variable group. select_cor_list is funtion for selecting highly correlated variable list.

Usage

get_correlation_group(cor_mat, p = 0.8)

select_cor_group(cor_vars)

select_cor_list(cor_vars_list)

Arguments

cor_mat
A correlation matrix of independent variables.

p
Threshold of correlation between features. Default is 0.7.

cor_vars
Correlated variables.

cor_vars_list
List of correlated variable

Value

A list of selected variables.

Examples

## Not run: 
cor_mat = cor(UCICreditCard[8:20],
use = "complete.obs", method = "spearman")
get_correlation_group(cor_mat, p = 0.6 )

## End(Not run)

[Package creditmodel version 1.1.2 Index]