get_ctree_rules {creditmodel}	R Documentation

Parse party ctree rules

Description

get_ctree_rules This function is used to parse party ctree rules and percentage of 1 under each rule.

Usage

get_ctree_rules(tree_fit = NULL, train_dat = NULL, target = NULL,
  test_dat = NULL, tree_control = list(p = 0.05, cp = 0.0001, xval = 1,
  maxdepth = 10), seed = 46)

Arguments

tree_fit
A tree model object.

train_dat
A data.frame of train.

target
The name of target variable.

test_dat
A data.frame of test.

tree_control
the list of parameters to control cutting initial breaks by decision tree.

seed
Random number seed. Default is 46.

Value

A data frame with tree rules and 1 percent under each rule.

Examples

train_test <- train_test_split(UCICreditCard, split_type = "Random", prop = 0.8, save_data = FALSE)
dat_train = train_test$train
dat_test = train_test$test
dat_train$default.payment.next.month = as.numeric(dat_train$default.payment.next.month)
get_ctree_rules(tree_fit = NULL, train_dat = dat_train[, 8:26],
target ="default.payment.next.month", test_dat = dat_test)

[Package creditmodel version 1.1.2 Index]