get_plots {creditmodel}	R Documentation

Plot Independent Variables

Description

plot_vars is used for independent variables with target variable ploting. get_plots can loop through plots for all specified independent variables.

Usage

get_plots(dat_train, dat_test = NULL, x_list = NULL, target = NULL,
  ex_cols = NULL, breaks_list = NULL, pos_flag = NULL,
  occur_time = NULL, equal_bins = FALSE, best = TRUE, g = 20,
  tree_control = NULL, bins_control = NULL, plot_show = TRUE,
  save_data = FALSE, file_name = NULL, parallel = FALSE,
  g_width = 8, dir_path = tempdir())

plot_vars(dat_train, x, dat_test = NULL, target = "target",
  g_width = 8, breaks_list = NULL, pos_flag = list("1", 1, "bad",
  "positive"), occur_time = NULL, equal_bins = FALSE, best = TRUE,
  g = 20, tree_control = NULL, bins_control = NULL,
  plot_show = TRUE, save_data = FALSE, dir_path = tempdir())

Arguments

dat_train
A data.frame with independent variables and target variable.

dat_test
A data.frame of test data. Default is NULL.

x_list
Names of independent variables.

target
The name of target variable.

ex_cols
A list of excluded variables. Regular expressions can also be used to match variable names. Default is NULL.

breaks_list
A table containing a list of splitting points for each independent variable. Default is NULL.

pos_flag
Value of positive class, Default is "1".

occur_time
The name of the variable that represents the time at which each observation takes place.

equal_bins
Logical, generates initial breaks for equal frequency binning.

best
Logical, merge initial breaks to get optimal breaks for binning.

g
Number of initial breakpoints for equal frequency binning.

tree_control
Parameters of using Decision Tree to segment initial breaks. See detials: get_tree_breaks

bins_control
Parameters used to control binning. See detials: select_best_class, select_best_breaks

plot_show
Logical, show model performance in current graphic device. Default is FALSE.

save_data
Logical, save results in locally specified folder. Default is FALSE.

file_name
The name for periodically saved data file. Default is NULL.

parallel
Logical, parallel computing. Default is FALSE.

g_width
The width of graphs.

dir_path
The path for periodically saved graphic files.

x
The name of an independent variable.

Examples

train_test <- train_test_split(UCICreditCard[1:1000,], split_type = "Random",
 prop = 0.8, save_data = FALSE)
dat_train = train_test$train
dat_test = train_test$test
get_plots(dat_train[, c(8, 26)], dat_test = dat_test[, c(8, 26)],
target = "default.payment.next.month")

[Package creditmodel version 1.1.2 Index]