get_sim_sign_lambda {creditmodel}	R Documentation

get_sim_sign_lambda get_sim_sign_lambda is for get Best lambda required in lasso_filter. This function required in lasso_filter

Description

get_sim_sign_lambda get_sim_sign_lambda is for get Best lambda required in lasso_filter. This function required in lasso_filter

Usage

get_sim_sign_lambda(lasso_model, sim_sign = "negtive")

Arguments

lasso_model
A lasso model genereted by glmnet.

sim_sign
Default is "negtive". This is related to pos_plag. If pos_flag equals 1 or 1, the value must be set to negetive. If pos_flag equals 0 or 0, the value must be set to positive.

Details

lambda.sim_sign give the model with the same positive or negetive coefficients of all variables.

Value

Lanmbda value

[Package creditmodel version 1.1.2 Index]