local_outlier_factor {creditmodel}	R Documentation

local_outlier_factor local_outlier_factor is function for calculating the lof factor for a data set using knn This function is not intended to be used by end user.

Description

local_outlier_factor local_outlier_factor is function for calculating the lof factor for a data set using knn This function is not intended to be used by end user.

Usage

local_outlier_factor(dat, k = 10)

Arguments

dat
A data.frame contained only predict variables.

k
Number of neighbors for LOF.Default is 10.

[Package creditmodel version 1.1.2 Index]