lr_params {creditmodel}	R Documentation

Logistic Regression & Scorecard Parameters

Description

lr_params is the list of parameters to train a LR model or Scorecard using in training_model.

Usage

lr_params(tree_control = list(p = 0.02, cp = 0.00000001, xval = 5,
  maxdepth = 10), bins_control = list(bins_num = 10, bins_pct = 0.05,
  b_chi = 0.02, b_odds = 0.1, b_psi = 0.03, b_or = 0.15, mono = 0.2,
  odds_psi = 0.15, kc = 1), best_lambda = "lambda.sim_sign",
  sp_values = NULL, forced_in = NULL, obsweight = c(1, 1),
  lasso = TRUE, vars_plot = TRUE, step_wise = TRUE,
  score_card = TRUE, cor_p = 0.8, iv_i = 0.02, psi_i = 0.1, ...)

Arguments

tree_control
the list of parameters to control cutting initial breaks by decision tree. See details at: get_tree_breaks

bins_control
the list of parameters to control merging initial breaks. See details at: select_best_breaks,select_best_class

best_lambda
Metheds of best lanmbda stardards using to filter variables by LASSO. There are 3 methods: ("lambda.auc", "lambda.ks", "lambda.sim_sign") . Default is "lambda.auc".

sp_values
Vaules will be in separate bins.e.g. list(-1, "Missing") means that -1 & Missing as special values.Default is NULL.

forced_in
Names of forced input variables. Default is NULL.

obsweight
An optional vector of 'prior weights' to be used in the fitting process. Should be NULL or a numeric vector. If you oversample or cluster diffrent datasets to training the LR model, you need to set this parameter to ensure that the probability of logistic regression output is the same as that before oversampling or segmentation. e.g.:There are 10,000 0 obs and 500 1 obs before oversampling or under-sampling, 5,000 0 obs and 3,000 1 obs after oversampling. Then this parameter should be set to c(10000/5000, 500/3000). Default is NULL..

lasso
Logical, if TRUE, variables filtering by LASSO. Default is TRUE.

vars_plot
Logical, if TRUE, plot distribution and correlation of input variables . Default is TRUE.

step_wise
Logical, stepwise method. Default is TRUE.

score_card
Logical, transfer woe to a standard scorecard. If TRUE, Output scorecard, and score prediction, otherwise output probability. Default is TRUE.

cor_p
The maximum threshold of correlation. Default: 0.8.

iv_i
The minimum threshold of IV. 0.01 to 0.1 usually work. Default: 0.02

psi_i
The maximum threshold of PSI. 0.1 to 0.3 usually work. Default: 0.1.

...
Other parameters

Value

A list of parameters.

See Also

training_model, xgb_params, gbm_params, rf_params

[Package creditmodel version 1.1.2 Index]