outliers_detection {creditmodel}	R Documentation

Outliers Detection outliers_detection is for outliers detecting using Kmeans and Local Outlier Factor (lof)

Description

Outliers Detection outliers_detection is for outliers detecting using Kmeans and Local Outlier Factor (lof)

Usage

outliers_detection(dat, x, kc = 3, kn = 5)

Arguments

dat
A data.frame with independent variables.

x
The name of variable to process.

kc
Number of clustering centers for Kmeans

kn
Number of neighbors for LOF.

Value

Outliers of each variable.

[Package creditmodel version 1.1.2 Index]