reduce_high_cor {creditmodel}	R Documentation

Compare the two highly correlated variables

Description

reduce_high_cor is function for comparing the two highly correlated variables, select a variable with the largest IV value.

Usage

reduce_high_cor(cor_mat, p = 0.9, x_list = NULL, com_list = NULL,
  retain = TRUE)

Arguments

cor_mat
A correlation matrix.

p
The threshold of high correlation.

x_list
Names of independent variables.

com_list
A data.frame with important values of each variable. eg : IV_list.

retain
Logical, output selected variables, if FALSE, output filtered variables.

Value

A list of selected variables.

[Package creditmodel version 1.1.2 Index]