require_packages {creditmodel}	R Documentation

Packages required and intallment

Description

require_packages is function for librarying required packages and installing missing packages if needed.

Usage

require_packages(pkg)

Arguments

pkg
A list or vector of names of required packages.

Value

packages installed and library.

Examples

## Not run: 
require_packages(c("dplyr","ggplot2"))

## End(Not run)

[Package creditmodel version 1.1.2 Index]