rf_params {creditmodel}	R Documentation

Random Forest Parameters

Description

rf_params is the list of parameters to train a Random Forest using in training_model.

Usage

rf_params(ntree = 100, nodesize = 30, samp_rate = 0.5,
  tune_rf = FALSE, ...)

Arguments

ntree
Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times.

nodesize
Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time). Note that the default values are different for classification (1) and regression (5).

samp_rate
Percentage of sample to draw. Default is 0.2.

tune_rf
A logical.If TRUE, then tune Random Forest model.Default is FALSE.

...
Other parameters

Details

See details at : https://www.stat.berkeley.edu/~breiman/Using_random_forests_V3.1.pdf

Value

A list of parameters.

See Also

training_model, lr_params, gbm_params, xgb_params

[Package creditmodel version 1.1.2 Index]