vintage_function {creditmodel}	R Documentation

vintage_function vintage_function is for vintage analysis.

Description

This function is not intended to be used by end user.

Usage

vintage_function(dat, obs_id = NULL, occur_time = NULL, MOB = NULL,
  period = "monthly", status = NULL, amount = NULL, by_out = "cnt",
  start_date = NULL, end_date = NULL, dead_status = 30)

Arguments

dat
A data.frame contained id, occur_time, mob, status ...

obs_id
The name of ID of observations or key variable of data. Default is NULL.

occur_time
The name of the variable that represents the time at which each observation takes place.

MOB
Mobility of book

period
Period of event to analysiss. Default is "monthly"

status
Status of observations

amount
The name of variable representing amount. Default is NULL.

by_out
Output: amount (amt) or count (cnt)

start_date
The earliest occurrence time of observations.

end_date
The latest occurrence time of observations.

dead_status
Status of dead observations.

[Package creditmodel version 1.1.2 Index]