xgb_params {creditmodel}	R Documentation

Logistic Regression & Scorecard Parameters

Description

xgb_params is the list of parameters to train a LR model or Scorecard using in training_model.

Usage

xgb_params(nrounds = 1000, params = list(max.depth = 6, eta = 0.1,
  min_child_weight = 1, subsample = 1, colsample_bytree = 1, gamma = 0,
  max_delta_step = 0, eval_metric = "auc", objective = "binary:logistic"),
  early_stopping_rounds = 100, ...)

Arguments

nrounds
Max number of boosting iterations.

params
A list contains parameters of xgboost.The complete list of parameters is available at: http://xgboost.readthedocs.io/en/latest/parameter.html

early_stopping_rounds
If NULL, the early stopping function is not triggered. If set to an integer k, training with a validation set will stop if the performance doesn't improve for k rounds.

...
Other parameters

Value

A list of parameters.

See Also

training_model, lr_params,gbm_params, rf_params

[Package creditmodel version 1.1.2 Index]